pyproject.toml
[tool.poetry]
name = "qa-guru-homework-17"
version = "0.1.0"
description = "API tests for lesson_17"
authors = ["Lev Savinkov"]
readme = "README.md"
packages = [{ include = "qa" }]

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.3"
requests = "^2.32.3"
pydantic = "^2.10.1"
allure-pytest = "^2.13.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--clean-alluredir --alluredir=allure-results"
testpaths = ["tests"]